ajout du modele compte avec restructuration du code, protection acces des pages du jeu, avancement affichage carte

+++ hegoa_2015/includes/class_page_djun_suppression_validation.php
@@ -1,6 +1,6 @@
 <?php
 // ======================================================================
-// Auteur : Donatien CELIA
+// Auteur : Donatien CELIA, Dominique Dehareng
 // Licence : CeCILL v2
 // ======================================================================
 
@@ -25,37 +25,42 @@ class PageDjunSuppressionValidation extends Page
     // - Affichage de la page
     public function Afficher()
     {
+		if(isset($_SESSION["compte"])&&!empty($_SESSION["compte"])){
+     	 // - gestion spécifique de la page
+      		$account_id           = $_SESSION['compte']->id;
+      		$djun_name            = $_SESSION['djun_choisi']->nom;
 
-      // - gestion spécifique de la page
-      $account_id           = $_SESSION['account_id'];
-      $djun_name            = $_SESSION['djun_choisi']->nom;
+      		// - On supprime le D'jun
+      		$sql  = "DELETE FROM \"libertribes\".\"AVATAR\"  WHERE compte_id = $account_id and avatar_nom ='$djun_name'";
 
-      // - On supprime le D'jun
-      $sql  = "DELETE FROM \"libertribes\".\"AVATAR\"  WHERE compte_id = $account_id and avatar_nom ='$djun_name'";
-
-      $result = $this->db_connexion->Requete( $sql );
-      if ( ! $result || pg_affected_rows($result)==0)
-      {
-        // - redirection vers la page djun_suppression avec un message d'erreur
-        header('Location: index.php?page=djun_suppression&erreur=1');
-        exit;
-      }
-		//  restructurer la variable de session avatar s'il y a plusieurs avatars
-		$avatars = array();
-		$i=0;
-		foreach($_SESSION['avatars'] as $avatar){
-			if($avatar->nom != $_SESSION['djun_choisi']->nom){
+      		$result = $this->db_connexion->Requete( $sql );
+      		if ( ! $result || pg_affected_rows($result)==0)
+      		{
+      		  // - redirection vers la page djun_suppression avec un message d'erreur
+      		  header('Location: index.php?page=djun_suppression&erreur=1');
+      		  exit;
+      		}
+			//  restructurer la variable de session avatar s'il y a plusieurs avatars
+			$avatars = array();
+			$i=0;
+			foreach($_SESSION['avatars'] as $avatar){
+				if($avatar->nom != $_SESSION['djun_choisi']->nom){
 				$avatars[$i] = $avatar;
 				$i++;
-			}
-		}	
-		unset($_SESSION['avatars']);	
-		$_SESSION['avatars'] = $avatars;
-		unset($_SESSION['djun_choisi']);
-		unset($_SESSION["avatar_name"]);
+				}
+			}	
+			unset($_SESSION['avatars']);	
+			$_SESSION['avatars'] = $avatars;
+			unset($_SESSION['djun_choisi']);
+			unset($_SESSION["avatar_name"]);
 		
-      // - redirection vers la page TDB
-      header('Location: index.php?page=tdb');
+      		// - redirection vers la page TDB
+      		header('Location: index.php?page=tdb');
+   			}
+   		else {
+			header('Location: index.php?page=connexion&erreur=3');
+			exit;
+		}
     }// - Fin de la fonction Afficher
 
 }// - Fin de la classe
